feat(cart): Add indicator dot to cart button
Now if there are items in the cart, there is a dot next to the cart icon.

The indicator dot takes its value from a localStorage variable, which
is read when the cart loads. This is kept in sync every time the checkout
is updated. This means that the dot can be shown before the checkout is
fetched.

Now closing can be done via the esc button as well. This is done via a
`keydown` event on the cart element, so will not work if user tabs to
the main content

Also, made a separate div to hold the icon buttons. It is a grid, so
gap is used for spacing between icons. And improved the cart page
design a little bit.