feat(i18n): Enable adding e.g. "incl. tax" after product prices
Product price is now translated via `i18n` files so it's possible to
modify how product prices are shown.

Previously if there were many different prices for a product, the output
was hardcoded as "from ...". Now this can be edited as well as an `i18n`
entry.

The translated and formatted price is shown for product variants, but
is not shown in the cart (cart still shows just the price).

Fix an unreleased cart bug where adding a variant without first opening
the cart would create a new cart with just the added variant.

Closes #1