fix: Properly load hero images
This makes hero images "lazy loaded", except if the hero is the first module on the page, in which case it's loaded immediately. Now lazy loading should also work as intended on Safari.

Also fixes hero element html validity (`noscript` is not allowed in `picture`), layout if Javascript is disabled, and adds 'sizes` attribute to `source` elements.

Closes #60