fix: Properly tell browsers what resolution image to download
All images should now have been updated with a "best guess" `sizes` attribute. (Except for deprecated modules, they use `100vw` which means they still download images that are too big.)

The `sizes` attribute is very important and tells the browser what size the image is expected to be and thus which resolution to download.

Now the `image` partial also creates a warning if images are missing the `sizes` attribute.

Closes #59