fix: Cart event now fired also on Baby Kit page
Previously the logic for the add to cart event was only loaded on product pages. This fix brings that logic to pages which have "buy buttons" as well.

Now using only the information from the event to fire add to carts for the different services. So no reliance on product page data.

Closes #68