feat: Collection descriptions no longer imported from Shopify
Importing collection descriptions from Shopify does not work well (see linked issue). There were a number of workarounds to make this possible, but now the decision has been made to remove this functionality.

In the future, please edit the collection pages as you like in Forestry. You can of course still have descriptions in Shopify for other purposes, but be aware that these will not be used on the site.

Closes #74