fix: Make "shop all" button when a product list shows less than the total number of products in the collection
So if the collection contains 50 product, but the product list module is limited to 10, e.g. on the front page, a button to view the entire collection is shown. This is a link to the collection page. The button is translatable and the button text can also be set on a per-module basis.

This commit also contains the following fixes:
- made the "sold out" tag translatable
- refactored product list css to be better organized

Closes #82