feat: Add custom attribute on checkout if specific cookie found
If a specifically named cookie is found, the cart will now set a custom attribute on the checkout corresponding to that cookie. The cookie format is `X-Checkout-Attr-[key]=[value]`, where the custom attribute `[key]` will be set to `[value]` when the checkout is created. Only one attribute is currently set.

This is currently used by the A8 integration, which sets the `X-Checkout-Attr-A8` cookie.

Closes #87