fix: Fix image grid popups not working
The popups are now shown based on the `:target` selector. This is how the id for the popups are determined:

- If the module has a heading, use that as the "module" id. Otherwise create a random string.
- Use the first line of the popup contents as the "image" id. If there are multiple images with the same first line (e.g. same heading), append the image index to the id.
- Combine the "module" and "image" ids into the final id for the popup.

While this should in practice always work, it creates a very unlikely possibility of duplicate popup ids. This happens if there are two image grids with the same heading and they both have popups with the same first line of content. Of course the random id created for modules with no heading might be the same as that of another module, but this is only an extremely theoretical possibility.

The bug was due to the global `.overlay` class being changed. Now refactored the image grid module and got rid of all global classes and logic.

Closes #101