fix: Fix filtering error
This was due to `collections` missing on the product level. Now added collections to the query and transforming the results into an array of collection handles.

This is now done in the main `write-products.ts` workflow. Added a workaround in the legacy mapper wrapper to delete the collections, since they're not in the format the legacy mapper expects.

Closes #109