fix: Use new checkout worker route for cart
We no longer load the Shopify checkout SDK as part of the cart functionality. Instead, the `/cart` endpoint is called with the relevant parameters.

This change has the following benefits:
- Better code reuse. All business logic related to checkout handling now only exists in the cart worker. This includes e.g. cookie handling.
- Server-side logging. The cart worker can now include logic to e.g. report conversions from AB tests without the need to implement client-side calls and logging endpoints.

This change means that the cart worker needs to be configured properly for each new site with a Shopify token etc. as well as a route!

All existing functionality has been ported to the new solution.