fix: Create AB test for Klarna payment plan banner
This is an ongoing A/B test for the "4 interest-free payments ..." Klarna banner. The experiment needs to be enabled for the sites separately, so contact a developer if you want to test this. The banner is currently only in English.

The experiment name is `exp-klarna-banner`. Intentionally leaving the related issue open for the duration of the test.

See #86