fix: Implement tabbed design for product page
As a part of copying the product page design, the social sharing buttons were also removed.

The tabbing is implemented with radio boxes. On mobile, the design is an "accordion" style, which is implemented with the `details` element. Heavy usage of `display: contents` in order to get the correct layout and input events.