fix: Correctly show tabs in non-Firefox browsers
This fixes a bug where at least Chrome did not correctly show the contents of the product page tabs.

This was due to the details element, where Chrome did not show the contents on non-open details elements even with `display: content` set on the details element. This was fixed by creating duplicate content for the tabs vs the accordion. This is pretty bad html, but at least this bug is now fixed, and the structure is now way easier from a development point of view.