feat: Visual "tags" for products in the product listing module
Tags to show in the product listing module can now be added by tagging products in Shopify. Products you want to show a tag for should be tagged in Shopify with `Site Tag:<tag>`, where "<tag>" is the textual tag you want to show (not including <>). Note that these tags are shown only on the product list module (not on the product page), and are not shown for sold out products (this becomes weird visually).

Also, added some padding to the product "cards" in the product list module.

Closes #141