fix: Add product impression tracking for Frosmo
There is now a new helper (get-product-impressions) that will get all the product pages of products shown in products modules on the page. This is used by the Frosmo pixel to list product impressions. Note that the product pages returned are the sub-pages of collections, so to get all the relevant information, call the get-product helper.

See #170