fix: Improve product image switching
Now the image in the product image carousel is changed whenever the main image for a variant changes (i.e. when the user changes the color). This now works better with the images from the media bank and does not require setting the image index in the frontmatter anymore. Furthermore, updated the carousel and thumbnails so that the thumbnails only depend on the carousel (not the other way around).

Added an event "variant-changed" that can be used in analytics.

Closes #81