fix: Add possibility to use product-color from sku as GA product id
This is now an opt-in feature to use product-color (e.g. 520232B-3882) as the product id for Google Analytics. The affected events are the product view event (`view_item`) and add to cart (`add_to_cart`) in Enhanced Ecommerce.

The SKU used for the product view id is that of the first available variant. If no variants are available the first variant is shown. This should mean that the variant that is shown on page load will always be used.

Opt-in is configured with the `gaSkuProductId` site parameter in the Hugo config. Note that products need to be imported with the new media bank -enabled importer for this to work.