feat: Allow linking to specific options (e.g. color) on product page
Link to a specific selection by adding the option name and value to the query string, i.e. `/products/overall?Color=Navy`. Both the option name and value should be how it is specified in Shopify and how it is visible on the product page itself. I.e. if "Color" is named "Farbe", you need to use the latter. Note that you can change any option in this way, e.g. link to a specific size as well.

Closes #191