fix: Rewrite link url handling to support subfolder setup
All links now work with a subfolder setup (e.g. www.reima.com/at). This means that now you only need to enter urls relative to the actual site in question. Links should thus now be entered as `/pages/page` instead of `/at/pages/page`. This works for all menu links, button links, etc. as well as links created in markdown content. (Sites without a subfolder setup should see no change in url handling.)

As a side effect of this, all urls on the site can now also be validated. This means that whenever you have a broken link, the system will warn you about this. Warnings are also created if links that should be relative (e.g. `/pages/page`) are created as absolute (e.g. `www.reima.com/at/pages/page`).

Developers should note that whenever a url is created for an anchor link or similar, the `helpers/get-link` partial should be used to resolve the full url. There is a convenience partial `helpers/get-link-module` that can be used when creating links in the context of a module. See source code for examples and documentation.

See #175