fix: Fix linking to color variant not always working
This was due to a race condition, where the `r-carousel` custom element might not have been defined when dispatching the `change` event. That event triggers a method available on the custom element `scrollToImage`, which obviously is not defined before the custom element is loaded and defined. Now waiting for `r-carousel` to be defined with `CustomElementRegistry.whenDefined()`, which should fix this issue.

Closes #220