fix: Fix announcement link not clickable bug
The announcements are technically on top of each other, which means only the last one was clickable (e.g. links did not work correctly). Now the inner announcement `div` (the one that animates and contains the content) has a `z-index` of `1`, so those are now always on top of the parent `divs` (the ones that are on top of each other).