wip: Create products module option to split color variants
New parameter for the `products` module `split_colors` will now render one `li` element per color instead of one per product. In this case, the color swatches will be hidden, and the color name printed under the product name instead. The link is also changed to point to the specific color in question.

This is implemented via a new helper partial that renders the `li` differently if it receives a specific color to render.