wip: Show all facets when returning to search results
The initial search on the results page is filtered according to the location hash. This means that the facets returned are also based on that filter. However, when going back to the search result page, we want to show all filters. This commit introduces a new search to get the facets without filters in order to render that into the filters element.

The following minor changes are also included:
- filter now renders facets that have selected filters as `open`
- filter element `render` now only takes an array of facets