wip(design): Update typography to match spec
Headings (h1-h6) and body text are now according to the spec. New css custom properties for setting font properties in one go are the `font` shorthands `--h?-font`. Letter spacing needs to be set separately, and thus the `--h?-letter-spacing` properties still exist. This for changing heading typography in CSS. If it can be set in HTML, you can use the `.g-h?` classes.